+++ .ebextensions/00-makeFiles.config
@@ -1,0 +1,12 @@
+files:
+  "/sbin/appstart" :
+    mode: "000755"
+    owner: webapp
+    group: webapp
+    content: |
+      #!/usr/bin/env bash
+      JAR_PATH=/var/app/current/application.jar
+      
+      # run app
+      killall java
+      java -Dfile.encoding=UTF-8 -jar $JAR_PATH
