--- src/test/java/challenge/nDaysChallenge/auth/AuthRepositoryTest.java
@@ -1,19 +1,0 @@
-package challenge.nDaysChallenge.auth;
-
-import challenge.nDaysChallenge.domain.member.Member;
-import challenge.nDaysChallenge.dto.request.member.MemberRequestDto;
-import org.springframework.test.context.transaction.BeforeTransaction;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
-
-public class AuthRepositoryTest {
-//
-//    @BeforeTransaction
-//    public void 회원가입(){
-//        MemberRequestDto memberRequestDto = new MemberRequestDto("abc@naver.com","123","aaa",1);
-//        Member member = memberRequestDto.toMember(passwordEncoder);
-//        given(memberRepository.save(any())).willReturn(member);
-//    }
-
-}
