+++ src/test/java/challenge/nDaysChallenge/dajim/DajimRepositoryTest.java
@@ -6,108 +6,126 @@ import challenge.nDaysChallenge.domain.dajim.Dajim;
 import challenge.nDaysChallenge.domain.dajim.Open;
 import challenge.nDaysChallenge.domain.room.*;
 import challenge.nDaysChallenge.dto.request.dajim.DajimUploadRequestDto;
+import challenge.nDaysChallenge.repository.dajim.DajimFeedRepository;
 import challenge.nDaysChallenge.repository.dajim.DajimRepository;
+import challenge.nDaysChallenge.repository.member.MemberRepository;
+import challenge.nDaysChallenge.repository.room.RoomRepository;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.annotation.Rollback;
+import org.springframework.test.context.transaction.BeforeTransaction;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
+import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 class DajimRepositoryTest {
 
-    @Autowired DajimRepository dajimRepository;
+    @Autowired
+    MemberRepository memberRepository;
 
-    @DisplayName("다짐 저장")
-    @Test
-    void saveDajim(){
-        //given
-        Member member = new Member("user@naver.com","12345","userN",1, Authority.ROLE_USER);
+    @Autowired
+    RoomRepository roomRepository;
 
-        GroupRoom room = new GroupRoom(member, "newRoom",new Period(LocalDate.now(),100L), Category.ROUTINE,4,"보상", 0, 0);
+    @Autowired
+    DajimRepository dajimRepository;
 
-        //when
-        DajimUploadRequestDto dajimUploadRequestDto = new DajimUploadRequestDto("다짐 내용", "PUBLIC");
-        Dajim newDajim = Dajim.builder()
-                .room(room)
-                .member(member)
-                .content(dajimUploadRequestDto.getContent())
-                .open(Open.valueOf(dajimUploadRequestDto.getOpen()))
+    @Autowired
+    DajimFeedRepository dajimFeedRepository;
+
+    Member savedMember1, savedMember2;
+    Room savedRoom;
+    Dajim savedDajim1, savedDajim2, savedDajim3;
+
+    @BeforeEach
+    void 멤버_룸_다짐_세팅(){
+        Member member1 = Member.builder()
+                .authority(Authority.ROLE_USER)
+                .id("user@naver.com")
+                .pw("1234")
+                .nickname("nick")
+                .image(3)
                 .build();
-        Dajim savedDajim = dajimRepository.save(newDajim);
 
-        checkDajimRoomUser(newDajim, room, member);
+        Member member2 = Member.builder()
+                .authority(Authority.ROLE_USER)
+                .id("user2@naver.com")
+                .pw("1234")
+                .nickname("nick2")
+                .image(3)
+                .build();
 
-        //then
-        assertThat(newDajim.getMember().getId()).isEqualTo(savedDajim.getMember().getId());
+        savedMember1 = memberRepository.save(member1);
+        savedMember2 = memberRepository.save(member2);
+
+        Room room = new SingleRoom("기상", new Period(LocalDate.now(),30L) , Category.ROUTINE,2, "", 0, 0);
+
+        savedRoom = roomRepository.save(room);
+
+        Dajim dajim1 = DajimUploadRequestDto.toDajim(room, member1, "내용1", "PUBLIC");
+        Dajim dajim2 = DajimUploadRequestDto.toDajim(room, member2, "내용2", "PUBLIC");
+        Dajim dajim3 = DajimUploadRequestDto.toDajim(room, member2, "내용3", "PUBLIC");
+
+        savedDajim1 = dajimRepository.save(dajim1);
+        savedDajim2 = dajimRepository.save(dajim2);
+        savedDajim3 = dajimRepository.save(dajim3);
     }
 
-    private static void checkDajimRoomUser(Dajim dajim, Room room, Member member){
-        if (dajim.getRoom()!=room || dajim.getMember()!=member){
-            throw new RuntimeException("다짐에 대한 권한이 없습니다.");
-        }
+    @Test
+    void 챌린지룸_다짐_조회(){
+        List<Dajim> dajims = dajimRepository.findAllByRoomNumber(savedDajim1.getRoom().getNumber())
+                .orElseThrow(()->new RuntimeException("해당 룸에서 다짐을 조회할 수 없습니다."));
+
+        System.out.println(dajims.stream().map(dajim -> dajim.getContent())
+                                            .collect(Collectors.toList()));
+
+        assertThat(dajims.size()).isEqualTo(3);}
+
+    @Test
+    void 다짐번호로_다짐_조회(){
+        Dajim dajim1 = dajimRepository.findByNumber(savedDajim1.getNumber())
+                .orElseThrow(()->new RuntimeException("다짐이 없습니다"));
+        Dajim dajim2 = dajimRepository.findByNumber(savedDajim2.getNumber())
+                .orElseThrow(()->new RuntimeException("다짐이 없습니다"));
+        Dajim dajim3 = dajimRepository.findByNumber(savedDajim3.getNumber())
+                .orElseThrow(()->new RuntimeException("다짐이 없습니다"));
+
+        assertThat(dajim1.getMember().getNumber()).isEqualTo(savedDajim1.getMember().getNumber());
+        assertThat(dajim2.getMember().getNumber()).isEqualTo(savedDajim2.getMember().getNumber());
+        assertThat(dajim3.getMember().getNumber()).isEqualTo(savedDajim3.getMember().getNumber());
     }
 
-    @DisplayName("다짐 수정")
     @Test
-    void modifyDajim(){
-        //given
-        Member member = new Member("user@naver.com","12345","userN",1, Authority.ROLE_USER);
+    void 특정_멤버의_다짐_전체_조회(){
+        List<Dajim> nickDajims = dajimRepository.findAllByMemberNickname("nick2")
+                .orElseThrow(()->new RuntimeException("해당 멤버에게 다짐이 없습니다."));
 
-        GroupRoom room = new GroupRoom(member, "newRoom",new Period(LocalDate.now(),100L), Category.ROUTINE,4,"보상", 0, 0);
-        DajimUploadRequestDto dajimUploadRequestDto = new DajimUploadRequestDto("다짐 내용", "PUBLIC");
-        Dajim dajim = Dajim.builder()
-                .room(room)
-                .member(member)
-                .content(dajimUploadRequestDto.getContent())
-                .open(Open.valueOf(dajimUploadRequestDto.getOpen()))
-                .build();
-        dajimRepository.save(dajim);
+        assertThat(nickDajims.size()).isEqualTo(2);
 
-        //when
-        Dajim newdajim = dajimRepository.findByDajimNumber(1L).orElseThrow(()->new RuntimeException("현재 다짐을 찾을 수 없습니다."));
-        Dajim updatedDajim = newdajim.update(Open.PRIVATE, "새 다짐");
-
-        //then
-        assertThat(updatedDajim.getContent()).isEqualTo("새 다짐");
+        System.out.println(nickDajims.stream().map(dajim ->
+                        dajim.getContent())
+                .collect(Collectors.toList()));
     }
 
-    @DisplayName("전체 룸멤버 다짐 조회")
     @Test
-    void viewDajims(){
-        //given
-        Member member = new Member("user@naver.com","12345","userN",1, Authority.ROLE_USER);
-        Member member2 = new Member("user2@naver.com","12345","userN2",1, Authority.ROLE_USER);
-        GroupRoom room = new GroupRoom(member, "newRoom",new Period(LocalDate.now(),100L), Category.ROUTINE,4, "", 0, 0);
+    void 다짐_피드_조회(){
+        List<Dajim> dajimFeed = dajimFeedRepository.findAllByOpen();
 
-        DajimUploadRequestDto dajimUploadRequestDto = new DajimUploadRequestDto("다짐 내용", "PUBLIC");
-
-        Dajim newDajim = Dajim.builder()
-                .room(room)
-                .member(member)
-                .content(dajimUploadRequestDto.getContent())
-                .open(Open.valueOf(dajimUploadRequestDto.getOpen()))
-                .build();
-        dajimRepository.save(newDajim);
-
-        Dajim newDajim2 = Dajim.builder()
-                .room(room)
-                .member(member2)
-                .content(dajimUploadRequestDto.getContent())
-                .open(Open.valueOf(dajimUploadRequestDto.getOpen()))
-                .build();
-        dajimRepository.save(newDajim2);
-
-        //when
-        List<Dajim> dajimList = dajimRepository.findAllByRoomNumber(room.getNumber()).orElseThrow(()-> new RuntimeException("다짐을 확인할 수 없습니다."));
-
-        //then
-        assertThat(dajimList.size()).isEqualTo(2);
+        assertThat(dajimFeed.size()).isEqualTo(3);
     }
 
 }
